requirements.txt
tqdm==4.55.0
Flask==1.1.2
paystackapi==2.0.0
pymongo==3.11.2
qrcode==6.1
Telethon==1.18.2
python_telegram_bot==13.1
python-dotenv==0.15.0
telegram==0.0.1
dnspython==2.1.0
uvicorn==0.13.3